Added write-time option to taco cmd ln tool
I added a write-time flag to the taco tool for the user to store the times of computation into a file. This option will be used by scripts that collect data about the tool's performance.

In order for the option to output useful data, the time option must also be specified, and computation must take place.