tensor,module: add a utility to allow for debugging generated kernels
This commit adds a function `debugCompileSource` to a `Tensor` that
allows for the `Tensor` to use a kernel from a provided source file
instead of generating a new one. This allows developers to add
prints/assertions to TACO generated code to debug faster.

Inspired by Amalee's PR (#302), I would have found a command like this
very useful for debugging generated code.