*: add the ability to apply filters to individual tensor elements
This commit adds the ability to perform filtering operations (similar to
Numpy's boolean filtering). An example of this is below:

```
// Construct a filtering function that selects values != 1.
auto filter = [](ir::Expr var) {
  return ir::Neq::make(var, ir::Literal::make(1));
}
// Add together all elements in b and c that aren't equal to 1.
a(i) = (b(i) | filter) + (c(i) | filter);
```

Note that this functionality only works when we can assume that the zero
element is both an identity and an annihilator for the group that the
algebra is defined over.