codegen: compile generated code with debug symbols in debug builds
Relates to #373.

This commit adjusts the compilation command for generated code to
compile code with debug symbols when TACO is built in debug mode.
Conveniently, this allows for debuggers (like the integrated one in
CLion) to automatically display the generated source code when a
segfault occurs (along with other things the IDE supports).