index_notation: temporary fix for segfault with user defined funcs
This commit has a temporary fix for a segfault encountered when the
`isomorphic` function is applied onto `IndexExpr`s that contain `Func`
applications. The casted pointer in `include/taco/util/functions.h:11`
that is dereferenced is `NULL`. This temporary fix is needed so that
benchmarking code can execute back-to-back iterations.