*: support placement onto faces of processor grids
This commit includes code-generation and mapper implementation for
placing partitions onto faces of processor grids. The general
implementation idea is to iterate over the full processor grid in the
mapper, but map tasks only to points that are on the desired face of the
grid. This is achieved by communicating the placement grid dimensions
via arguments in the task that are read out by the mapper.