*: adjust transfers to declare partition bounds at higher level tasks
This commit adjusts the strategy for transfers to declare partition
bounds for a transfer at all tasks above the transfer. This effectively
gives the runtime system more information about what tasks are using
what pieces of data, and might avoid false all-to-all communication
patterns. For example, in the Cannon algorithm, blindly giving each node
the full B and C matrix causes the runtime to think that all nodes need
to track all other nodes. If we push the information that a node needs
only the chunk of B on its row at the top level task, then the runtime
can restrict communication to just those nodes. In general, giving more
information about partitions to the runtime should result in better
performance.