*: add ability to up front create partitions required for operations
This commit allows the lowerer to separate the declaration of partitions
from the computation that uses them. For dense tensors, this lets us
declare the necessary partitions (at the top level) for an operation once,
so that we don't need to create top level partitions for every operation.
This should lower the amount of metadata tracking that Legion has to
perform and should hopefully provide a good foundation for sparse
operations that have data dependent partitioning operations.